+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for ChibiHash
 
+## 0.2.0.0 -- 2024-12-02
+
+* Added V2 implementation of ChibiHash - v1 still the default
+* Module exports both v1 and v2, API is the same
+* Added example usage to README
+
 ## 0.1.0.0 -- 2024-11-27
 
 * First version.